refactor: move PublishDto to init method and rename it to ReactiveDto
