+++ imports/api/Users/users.publication.ts
@@ -2,19 +2,19 @@ import { ProfilesService } from '../Profiles/profiles.service';
 import { UserResponseDto } from './dtos/user-response.dto';
 import { UserRepository } from './user.repository';
 import { Publication } from '/imports/common/decorators/publication.decorator';
-import { PublishDto } from '/imports/common/decorators/publish-dto.decorator';
+import { ReactiveDto } from '/imports/common/decorators/reactive-dto.decorator';
 import { BasePublication } from '/imports/common/publications/base.publication';
 
 @Publication('users')
-@PublishDto(UserResponseDto)
 export class UsersPublication extends BasePublication {
 	constructor(
 		private readonly userRepository: UserRepository = new UserRepository(),
-		private readonly profilesService: ProfilesService = new ProfilesService(null as any) // UserService not needed for getStaticProfilesForExternalUsers()
+		private readonly profilesService: ProfilesService = new ProfilesService(null as any)
 	) {
 		super();
 	}
 
+	@ReactiveDto(UserResponseDto)
 	init() {
 		const selector = { 'profile.profile': { $nin: this.profilesService.getStaticProfilesForExternalUsers() } };
 		return this.userRepository.findAll(selector, {
